Add files via upload
Creation of WhatsApp group (adding members, assigning admin, and sending a message) through Browser Automation. 

The virtual environment whatsappgroup is enclosed. Selenium and python are used to code the main script script.py. 

Please change the username and path based on the platform you're using. This will help config.py to work smoothly. 

In case the program quits unexpectedly, please check your internet connection. Since the task is to automate the browser, the program
can be subject to network fluctuations because of the browser in use. 

Please do report any issues and queries.